Melhorias Week Card: Security, Endpoints

+++ 12WeekMethod/src/main/java/com/mundim/WeekMethod/service/WeekCardService.java
@@ -1,57 +1,94 @@
 package com.mundim.WeekMethod.service;
 
 import com.mundim.WeekMethod.dto.WeekCardDTO;
+import com.mundim.WeekMethod.dto.update.UpdateWeekCardDTO;
+import com.mundim.WeekMethod.entity.User;
 import com.mundim.WeekMethod.entity.WeekCard;
 import com.mundim.WeekMethod.exception.BadRequestException;
+import com.mundim.WeekMethod.exception.UnauthorizedRequestException;
 import com.mundim.WeekMethod.repository.WeekCardRepository;
+import com.mundim.WeekMethod.security.AuthenticationService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class WeekCardService {
 
-    private WeekCardRepository weekcardRepository;
+    private final WeekCardRepository weekcardRepository;
+    private final UserService userService;
+    private final AuthenticationService authenticationService;
 
     @Autowired
-    public WeekCardService(WeekCardRepository weekcardRepository){
+    public WeekCardService(WeekCardRepository weekcardRepository,
+                           UserService userService,
+                           AuthenticationService authenticationService) {
         this.weekcardRepository = weekcardRepository;
+        this.userService = userService;
+        this.authenticationService = authenticationService;
     }
 
-    public WeekCard createWeekCard(WeekCardDTO weekCardDTO){
-        return weekcardRepository.save(new WeekCard(weekCardDTO));
+    public WeekCard createWeekCardForLoggedUser(WeekCardDTO weekCardDTO) {
+        Long userId = authenticationService.findUserByBearer().getId();
+        return weekcardRepository.save(new WeekCard(weekCardDTO, userId));
     }
 
-    public WeekCard findWeekCardById(Long weekCardId){
+    public List<WeekCard> findWeekCardsForLoggedUser() {
+        Long userId = authenticationService.findUserByBearer().getId();
+        return weekcardRepository.findWeekCardByUserId(userId);
+    }
+
+    public WeekCard findWeekCardById(Long weekCardId) {
+        verifyUserAuthorizationForWeekCard(weekCardId);
         return weekcardRepository.findById(weekCardId)
                 .orElseThrow(() -> new BadRequestException("Week Card not found"));
     }
 
-    public WeekCard updateWeekCardById(WeekCardDTO weekCardDTO, Long weekCardId){
+    public WeekCard updateWeekCardById(UpdateWeekCardDTO dto, Long weekCardId) {
         WeekCard weekCard = findWeekCardById(weekCardId);
-        if(weekCardDTO.userId() != null) weekCard.setUserId(weekCardDTO.userId());
-        if(weekCardDTO.description() != null) weekCard.setDescription(weekCardDTO.description());
-        if(weekCardDTO.weekStartDate() != null) weekCard.setWeekStartDate(weekCardDTO.weekStartDate());
-        if(weekCardDTO.weekEndDate() != null) weekCard.setWeekEndDate(weekCardDTO.weekEndDate());
-        if(weekCardDTO.notes() != null) weekCard.setNotes(weekCardDTO.notes());
+        verifyUserAuthorizationForWeekCard(weekCardId);
+
+        if (dto.description() != null) weekCard.setDescription(dto.description());
+        if (dto.weekStartDate() != null) {
+            weekCard.setWeekStartDate(dto.weekStartDate());
+            weekCard.setWeekEndDate(weekCard.getWeekStartDate().plusDays(7));
+        }
+        if (dto.notes() != null) weekCard.setNotes(dto.notes());
+
         return weekcardRepository.save(weekCard);
     }
 
-    public WeekCard deleteWeekCardById(Long weekCardId){
+    public WeekCard deleteWeekCardById(Long weekCardId) {
+        verifyUserAuthorizationForWeekCard(weekCardId);
         WeekCard weekCard = findWeekCardById(weekCardId);
         weekcardRepository.deleteById(weekCardId);
         return weekCard;
     }
 
-    public WeekCard addTaskToWeekCard(Long taskId, Long weekCardId){
+    public WeekCard addTaskToWeekCard(Long taskId, Long weekCardId) {
+        verifyUserAuthorizationForWeekCard(weekCardId);
         WeekCard weekCard = findWeekCardById(weekCardId);
         weekCard.getWeekTasksIds().add(taskId);
         return weekcardRepository.save(weekCard);
     }
 
-    public WeekCard removeTaskFromWeekCard(Long taskId, Long weekCardId){
+    public WeekCard removeTaskFromWeekCard(Long taskId, Long weekCardId) {
+        verifyUserAuthorizationForWeekCard(weekCardId);
         WeekCard weekCard = findWeekCardById(weekCardId);
         weekCard.getWeekTasksIds().remove(taskId);
         return weekcardRepository.save(weekCard);
+    }
+
+    private void verifyUserAuthorizationForWeekCard(Long weekCardId) {
+        User user = authenticationService.findUserByBearer();
+        WeekCard weekCard = weekcardRepository.findById(weekCardId)
+                .orElseThrow(() -> new BadRequestException("Week Card not found"));
+
+        if (!weekCard.getUserId().equals(user.getId()) &&
+                !user.getAuthorities().contains(new SimpleGrantedAuthority("ROLE_ADMIN")))
+            throw new UnauthorizedRequestException("Unauthorized action");
     }
 
 }
